fix(client): port should be int16

+++ client/client.go
@@ -26,7 +26,7 @@ type HttpTransport func(rt http.RoundTripper) http.RoundTripper
 type Client struct {
 	Protocol              string
 	Host                  string
-	Port                  int8
+	Port                  int16
 	Timeout               time.Duration
 	RequestTransformerMgr *httptransport.RequestTransformerMgr
 	HttpTransports        []HttpTransport
